Add support for Docker (#1)
* Added a Dockerfile for ubuntu bionic (18.04)

* Create multiarch build with GH Actions

We push a test tag first

* Proper tag name instead of deriving from command

* Disable NVENC completely

We disable this for now to check that ARM works. If it does we will need to have 2 different images, one not arm compatible

* We will disable x265 build as well as it seems to work that way

We will add it back at a later stage

* Update Dockerfile to support multiarch

We will disable NVIDIA NVENC encoder and H.265 encoding for all arm based systems until further notice.

On x86 it will be fine and we can go ahead and enable both. We will also install the nvidia library on the x86 image

* Forgot that we actually changed the filename haha

* Optimize the conditional check and the docker folder

We split the docker folder up by their "tags" so as to allow for future upgrades while keeping old builds
We have also streamlined the conditional checks to be a single line to optimize the flow

* Added removing of apt cache after installing packages to shrink package size

* Fixed dependency installation

* Ready for merge

Removed force checkout to 1.3.x tag as it has been resolved
Updated GitHub Actions script to allow building for both latest and bionic builds

* Added initial Focal support

Right now Focal for arm is broken af due to ubuntu, so we are only providing the x86 build for now. Going forward we will provide a proper full version as well

* Added Alpine Linux support and Focal x86 changes

Focal x86
- We now install meson from apt itself as it is properly updated in Focal

Alpine
- Added new build for Alpine
- Also added a shell script to build or install numactl based on arch (arm requires build)

* Disabled arm build for Alpine

Alpine will only work for x86 hence we will compile only that as well.
We may work on it in the future though

* Added command prompt for default execution to show version

For some reason arm is showing format exec error on rpi so...

* Pushing another workflow to test in particular docker building with v3 buildx action

Trying to see if we can get the rpi build running

* Update variables

* Lets just build armv6 as well in case we actually uses it O.o

Not sure if v6 will build tho heh

* Latest buildx setup is 3.2.0

* Update main docker file for this fact

We added armv6 for now but we will check the builds to make sure we can actually build a armv6 version